Move single-instance handling to its own file. Use a named mutex for ServiceWire.
