+++ Program.cs
@@ -27,16 +27,13 @@ of the authors and should not be interpreted as representing official policies,
 either expressed or implied, of the FreeBSD Project.
 */
 
-// For communication with the main instance. Use ServiceWire or just pipes.
-//#define USE_SERVICEWIRE
-
 using System;
 using System.Globalization;
 using System.IO;
 using System.Linq;
+using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Threading;
-using System.Threading.Tasks;
 using MatterHackers.Agg;
 using MatterHackers.Agg.Platform;
 using MatterHackers.Agg.UI;
@@ -47,14 +44,6 @@ using MatterHackers.SerialPortCommunication.FrostedSerial;
 using Microsoft.Extensions.Configuration;
 using Mindscape.Raygun4Net;
 using SQLiteWin32;
-
-#if USE_SERVICEWIRE
-using ServiceWire;
-#else
-using System.IO.Pipes;
-using System.Text;
-using System.Xml.Serialization;
-#endif
 
 namespace MatterHackers.MatterControl
 {
@@ -82,7 +71,7 @@ namespace MatterHackers.MatterControl
 
 		private static RaygunClient _raygunClient;
 
-		private const string ServicePipeName = "MatterControlMainInstance";
+		
 
 		[DllImport("Shcore.dll")]
 		static extern int SetProcessDpiAwareness(int PROCESS_DPI_AWARENESS);
@@ -149,6 +138,16 @@ namespace MatterHackers.MatterControl
 			}
 #endif
 
+			// If running in the build directory, use the StaticData at the project root.
+			{
+				var mainOutputDirectoryAttribute = MainOutputDirectoryAttribute.GetFromProgramAssembly();
+				var executableDirectory = new DirectoryInfo(AppDomain.CurrentDomain.BaseDirectory);
+				var buildOutputDirectory = new DirectoryInfo(Path.Combine(mainOutputDirectoryAttribute.ProjectRoot, mainOutputDirectoryAttribute.MainOutputDirectory));
+				// NOTE: No good way to determine whether two paths refer to the same object without pinvoke? So, just ignore case in all cases.
+				if (executableDirectory.FullName.Equals(buildOutputDirectory.FullName, StringComparison.OrdinalIgnoreCase))
+					StaticData.OverrideRootPath(Path.Combine(mainOutputDirectoryAttribute.ProjectRoot, "StaticData"));
+			}
+
 			// Set the global culture for the app, current thread and all new threads
 			CultureInfo.DefaultThreadCurrentCulture = CultureInfo.InvariantCulture;
 			CultureInfo.DefaultThreadCurrentUICulture = CultureInfo.InvariantCulture;
@@ -176,138 +175,16 @@ namespace MatterHackers.MatterControl
 			};
 
 			// If MatterControl isn't running and valid files were shelled, schedule a StartupAction to open the files after load
-			var shellFiles = args.Where(f => File.Exists(f) && ApplicationController.ShellFileExtensions.Contains(Path.GetExtension(f).ToLower()));
+			var shellFiles = args.Where(f => File.Exists(f) && ApplicationController.ShellFileExtensions.Contains(Path.GetExtension(f).ToLower())).ToArray();
 
 			// Single instance handling.
 #if !DEBUG
-#if USE_SERVICEWIRE
-			try
+			if (!LocalService.TryStartServer())
 			{
-				using (var client = new ServiceWire.NamedPipes.NpClient<IMainService>(new ServiceWire.NamedPipes.NpEndPoint(ServicePipeName)))
-				{
-					if (client.IsConnected)
-					{
-						// and at least one argument is an acceptable shell file extension
-						if (shellFiles.Any())
-						{
-							// notify the running instance of the event
-							client.Proxy.ShellOpenFile(shellFiles.ToArray());
-						}
-
-						System.Threading.Thread.Sleep(1000);
-
-						// Finally, close the process spawned by Explorer.exe
-						return;
-					}
-				}
-			}
-			catch (Exception)
-			{
-			}
-
-			var host = new ServiceWire.NamedPipes.NpHost(ServicePipeName, new ServiceWireLogger());
-			host.AddService<IMainService>(new LocalService());
-			host.Open();
-#else
-			NamedPipeServerStream pipeServer = null;
-			try
-			{
-				pipeServer = new NamedPipeServerStream(ServicePipeName, PipeDirection.InOut, 1, PipeTransmissionMode.Message, PipeOptions.CurrentUserOnly);
-			}
-			catch (IOException)
-			{
-				// Server pipe already exists.
-			}
-
-			static string readPipeMessage(PipeStream pipe)
-			{
-				MemoryStream ms = new MemoryStream();
-				using var cancellation = new CancellationTokenSource();
-				byte[] buffer = new byte[1024];
-				do
-				{
-					var task = pipe.ReadAsync(buffer, 0, buffer.Length, cancellation.Token);
-					cancellation.CancelAfter(1000);
-					task.Wait();
-					ms.Write(buffer, 0, task.Result);
-					if (task.Result <= 0)
-						break;
-				} while (!pipe.IsMessageComplete);
-
-				return Encoding.Unicode.GetString(ms.ToArray());
-			}
-
-			if (pipeServer != null)
-			{
-				new Task(() =>
-				{
-					try
-					{
-						var localService = new LocalService();
-
-						for (; ; )
-						{
-							pipeServer.WaitForConnection();
-
-							try
-							{
-								string str = readPipeMessage(pipeServer);
-
-								var serializer = new XmlSerializer(typeof(InstancePipeMessage));
-								localService.ShellOpenFile(((InstancePipeMessage)serializer.Deserialize(new StringReader(str))).Paths);
-
-								using var cancellation = new CancellationTokenSource();
-								var task = pipeServer.WriteAsync(Encoding.Unicode.GetBytes("ok"), cancellation.Token).AsTask();
-								cancellation.CancelAfter(1000);
-								task.Wait();
-							}
-							catch (Exception)
-							{
-							}
-
-							// NamedPipeServerStream can only handle one client ever. Need a new server pipe. ServiceWire does the same thing.
-							// So here, there is a time where there is no server pipe. Another instance could become the main instance.
-							pipeServer.Dispose();
-							pipeServer = new NamedPipeServerStream(ServicePipeName, PipeDirection.InOut, 1, PipeTransmissionMode.Message, PipeOptions.CurrentUserOnly);
-						}
-					}
-					catch (Exception ex) // TimeoutException or IOException
-					{
-						//System.Windows.Forms.MessageBox.Show(ex.ToString());
-						System.Diagnostics.Trace.WriteLine("Main instance pipe server died: " + ex.ToString());
-					}
-
-					pipeServer.Dispose();
-					pipeServer = null;
-				}).Start();
-			}
-			else if (shellFiles.Any())
-			{
-				try
-				{
-					using var pipeClient = new NamedPipeClientStream(".", ServicePipeName, PipeDirection.InOut, PipeOptions.CurrentUserOnly);
-					pipeClient.Connect(1000);
-					pipeClient.ReadMode = PipeTransmissionMode.Message;
-
-					StringBuilder sb = new();
-					using (var writer = new StringWriter(sb))
-						new XmlSerializer(typeof(InstancePipeMessage)).Serialize(writer, new InstancePipeMessage { Paths = shellFiles.ToArray() });
-
-					using var cancellation = new CancellationTokenSource();
-					var task = pipeClient.WriteAsync(Encoding.Unicode.GetBytes(sb.ToString()), cancellation.Token).AsTask();
-					cancellation.CancelAfter(1000);
-					task.Wait();
-					if (task.IsCompletedSuccessfully && readPipeMessage(pipeClient).Trim() == "ok")
-						return;
-				}
-				catch (Exception ex) // TimeoutException or IOException
-				{
-					//System.Windows.Forms.MessageBox.Show(ex.ToString());
-					System.Diagnostics.Trace.WriteLine("Instance pipe client died: " + ex.ToString());
-				}
+				if (shellFiles.Any() && LocalService.TrySendToServer(shellFiles))
+					return;
 			}
 #endif
-#endif // Single instance handling
 
 			if (shellFiles.Any())
 			{
@@ -424,8 +301,6 @@ namespace MatterHackers.MatterControl
 			rootSystemWindow.ShowAsSystemWindow();
 		}
 
-		private static readonly object locker = new object();
-
 		static void KeepAwake(bool keepAwake)
 		{
 			if (keepAwake)
@@ -438,97 +313,24 @@ namespace MatterHackers.MatterControl
 				SetThreadExecutionState(EXECUTION_STATE.ES_CONTINUOUS);
 			}
 		}
-
-		[Serializable]
-		public class LocalService : IMainService
-		{
-			public void ShellOpenFile(string[] files)
-			{
-				// If at least one argument is an acceptable shell file extension
-				var itemsToAdd = files.Where(f => File.Exists(f)
-					&& ApplicationController.ShellFileExtensions.Contains(Path.GetExtension(f).ToLower()));
-
-				if (itemsToAdd.Any())
-				{
-					lock (locker)
-					{
-						// Add each file
-						foreach (string file in itemsToAdd)
-						{
-							ApplicationController.Instance.ShellOpenFile(file);
-						}
-					}
-				}
-			}
-		}
-
-#if USE_SERVICEWIRE
-		private class ServiceWireLogger : ServiceWire.ILog
-		{
-			static private void Log(ServiceWire.LogLevel level, string formattedMessage, params object[] args)
-			{
-				// Handled as in https://github.com/tylerjensen/ServiceWire/blob/master/src/ServiceWire/Logger.cs
-
-				if (null == formattedMessage)
-					return;
-
-				if (level <= LogLevel.Warn)
-				{
-					string msg = (null != args && args.Length > 0)
-						? string.Format(formattedMessage, args)
-						: formattedMessage;
-
-					System.Diagnostics.Trace.WriteLine(msg);
-				}
-			}
-
-			void ILog.Debug(string formattedMessage, params object[] args)
-			{
-				Log(LogLevel.Debug, formattedMessage, args);
-			}
-
-			void ILog.Error(string formattedMessage, params object[] args)
-			{
-				Log(LogLevel.Error, formattedMessage, args);
-			}
-
-			void ILog.Fatal(string formattedMessage, params object[] args)
-			{
-				Log(LogLevel.Fatal, formattedMessage, args);
-			}
-
-			void ILog.Info(string formattedMessage, params object[] args)
-			{
-				Log(LogLevel.Info, formattedMessage, args);
-			}
-
-			void ILog.Warn(string formattedMessage, params object[] args)
-			{
-				Log(LogLevel.Warn, formattedMessage, args);
-			}
-		}
-#else
-		[Serializable]
-		public struct InstancePipeMessage
-		{
-			public string[] Paths;
-		}
-#endif
 	}
 
-	public interface IMainService
-	{
-		void ShellOpenFile(string[] files);
-	}
 
 	[AttributeUsage(AttributeTargets.Assembly)]
 	public class MainOutputDirectoryAttribute : Attribute
 	{
-		public readonly string MainOutputDirectory;
+		public readonly string MainOutputDirectory; // Relative to ProjectRoot.
+		public readonly string ProjectRoot; // Absolute
 
-		public MainOutputDirectoryAttribute(string path)
+		public MainOutputDirectoryAttribute(string path, string projectRoot)
 		{
 			MainOutputDirectory = path;
+			ProjectRoot = projectRoot;
+		}
+
+		public static MainOutputDirectoryAttribute GetFromProgramAssembly()
+		{
+			return typeof(MatterControl.Program).Assembly.GetCustomAttribute<MainOutputDirectoryAttribute>();
 		}
 	}
 }
