+++ Tests/MatterControl.Tests/MatterControl/MatterControlUtilities.cs
@@ -76,7 +76,7 @@ namespace MatterHackers.MatterControl.Tests.Automation
 		public static readonly string TestsPath = Path.Combine(RootPath, "Tests");
 		public static readonly string MainBinOutputPath = Path.Combine(RootPath,
 			// This attribute is generated by an MSBuild task.
-			typeof(MatterControl.Program).Assembly.GetCustomAttribute<MainOutputDirectoryAttribute>().MainOutputDirectory
+			MainOutputDirectoryAttribute.GetFromProgramAssembly().MainOutputDirectory
 			);
 
 		private static SystemWindow rootSystemWindow;
